requirements.txt
openai==1.69.0
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3